Modify student and teacher part. For student part, can view the last submission in a month, and could be edit if status is approved or pending. Add viewing uploaded history part. For teacher part, in participant details, you can edit student's working hours approval status and performance score. In wage payment, you can get the payment record in this month and mark it as paied to chage its paied status.
Note that the time navigation in these two parts is not send as parameter but filter for the receiving data (if I do not make any wrong understanding).

However, there must be many bugs in these two parts. For example, edit submission status in student part haven't tried corectlly yet,
and the wage payment in teacher part could not be clicked on to show. I could not figure out why. Hard for me.